code for debugging issue. We find the hessian in internal coordinate Hq is unreasonably large along certain dimensions.
This is shown if we look at g_q - g_q0 & Hq0 (q-q0).

When we use the mean function that is Taylor expansion to second order around the reference point q0, the fitting of GPR model of the potential energy surface goes wrong.

If we use the constant mean function, the fitting of potential & gradient is successful. However, fitting of hessian is again failed.
This indicates hessian is problematic in internal coordinate.