(1) put model, likelihood, train_inputs and  train_targets on cuda. (2) when load the model hyper-parameter for the gpr, we need to specify the map_location for the loaded tensor. now add map position.
(3) When converge the GPR, we need to increase the force error criterion if the current criterion does not align with the force uncertainty after we update the model. However, the variance of force is different when we predict result for only one bead or all beads. Therefore, we now change the code to increase uncertainty for the case that we predict all beads' uncertainty.